Some linting and UI for login
I created the UI for the frontend. It's a very simple page with a box in
the middle of the screen. We can improve this later on but at least the
backend won't be blocked now.

I also added a linter that fixed a lot of space and styl issues through
out our vue files. This is why all the component files are updated but
no new code is introduced.